+++ server/routes/account.router.js
@@ -439,29 +439,6 @@ router.post('/importCurrentJSON', rejectUnauthenticated, async (req, res) => {
   // }
 });
 
-/**
-* GET route to log status of a user's loop
-*/
-router.get('/debug', rejectUnauthenticated, async (req, res) => {
-  const userID = req.query.id;
-  if (req.user.admin) {
-    try {
-      const userInfo = cache.getSafeStorage(userID);
-      const userErrors = cache.getErrors(userID);
-      console.log('debug - full storage', userInfo);
-      console.log('errors', userErrors);
-      userInfo.userID !== null
-        ? res.send(userInfo).status(200)
-        : res.sendStatus(500);
-    } catch (err) {
-      console.log(err, 'problem debug route');
-      res.sendStatus(500)
-    }
-  } else {
-    console.log('error debug user route - not admin');
-    res.sendStatus(403);
-  }
-});
 
 /**
 * GET route to get user's errors from cache
